Test event image file upload
Cypress doesn't provide a built-in way to interact with file inputs.
We can work around this for now by manually setting the input's files property, using a DataTransfer object to get a FileList.
I'm not sure if the browser is supposed to allow JS code to create DataTransfer instances, so this might break in a future Chrome update.