Fix e-mail filter
organizer contact should be filtered out if the user accessing them is
not an admin. This filter needs to be applied to any endpoint that
allows non-authenticated access.

To get this working a bit of tech debt had to be payed down. It was the
expectation that we would return an EventDTO model over the wire rather
than the database model. This gives us space to normalize the format for
the api rather than except whatever format sequelize returns, for
example we currently flatten venues to venue since we only allow one
venue right now and we do some normalization of data/time values.

This transformation was not getting applied everywhere but our our
filter functions expected the transformed type.

This PR:

1. Makes sure all even endpoints expose the proper DTO
2. Applies organizer_contact filter to all endpoints that don't require
   authentication
3. Updates frontend code that wasn't expecting the venue field to
   already be set

Notes:

* If you hit an endpoint that does not require authentication but
you are authenticated as a site admin the filter will not be applied and
you will get back the organizer_contact.
* We don't bother invoking the filter at all on endpoints that require
  authentication as we don't have any non-admin user roles at the
  moment. If you are authenticated you are admin so even if we applied
  the filtering code it would see that your are an admin and doing
  nothing.