Run test workflows when workflow definitions change
We limit the test jobs to run only when files they cover have changed, to save on Github actions minutes. But that makes it hard to test changes to the workflows themselves.

Adding the workflow definition files to the paths list should trigger them if we change something in the definition, without requiring us to change unrelated code for testing.