Draft: Add pagination to verified events endpoint (#426)
Adds pagination support to events/verified

This endpoint returns all (not just current) verified events. Historically this endpoint has not worked because it returns too much data. We don't actually use it currently but it could be useful in the future and will likely be needed for many third party integrations.

By adding pagination this endpoint should now be usable.

    The default page is 1 and the default pageSize is 20, but these can be changed by the user.
    The pagination can also be paired with a pre-filter on a single or multiple tags
    Only verified events are returned

I have dropped support for the includes tag which let you specify if nested models like venue or date_times are included. To do the filtering and sorting we need to fetch them in all cases and given that are models are not super large they do not seem necessary.

Since no we are not actually leveraging this endpoint and it hasn't been working the risk of breaking changes is low. I would like to possibly generalize some of these changes to our other endpoints next.

Co-authored-by: Matthew Gidcomb <matthew.gidcomb@outlook.com>