Move Authenticated Endpoints (#437)
Moved the endpoints for getting all events and for getting non-verified endpoints to the authenticated controller. This is a bit of house keeping. The intent was that all endpoints that require authentication would be under a /authenticated route. This is for external communication and it also makes it so we can turn auth at the controller level rather than on each endpoint.

I also backfilled some tests