#356 Part I -- Move Image Upload API (#454)
I am preparing to implement a first pass of basic image per-processing which will make sure we do not store images that are too large. As an initial first step I have moved image processing out the web-portal and into the API-Server where we are supported by the more robust backend functionality that NextJs provieds.

This inadvertanly addresses issue #436. In the re-write I have not ported the bug :-)