Issue-356: Downsample Images -- Part II (#459)
This will check if the incoming image is greater than 1920 pixels wide, if so it resizes down to 1920. Regardless of width it will save the image from the incoming format to a compressed webp format. With some quick local testing I'm seeing images from my phone drop from 3-9megs down to a few hundred kb. They still look totally fine to my untrained eye even on my fancy super high ultra whatever we call it these monitor. This is especially true because even in the full event view the image is not allowed to fill the screen.

A better version of this would actually save at least 2 images in mind, for example we could save an even smaller version for the tiles. I am considering that change for a followup, but that will require migrations and more code changes. This version should go along way towards improving our standing with google and mobile load times using only small change