Issue #470: Update Node to lts/iron 20.x.x
Node 16 is end of life, it's time to move on. I've updated the .nvmrc
file, the readme files, the github actions configuration,  and the
container definitions to point to the latest node lts, which is 20,
code named iron.

Preliminary local testing indicates no issues. Assuming all goes well in
CI/CD we can move on to a quick staging test.