resolve flaky test
We need to sort on the first_start_time a second time after joining our page
back to the events table becuase the order that comes out of the merge is not
guaranteed to be sorted the same as in it was inside the common expression table

basically we were sorting the overall list correctly and guaranteeing each page would always have
the same things on it but that the things within each page would always be in the same order.

Applying this sort correctly revealed a second test which was wrongly assuming online resources, which
have no date_times would come at the end of the page when in fact they will always come at the beginnning
of the page.

To further tighten this up I added a secondary sort on created_at so that not only will online resources always
appear at the beginning of the page, but also in the same order

I left in place a new test based on data which produced durring randomized runs that originally highlighted this
problem