Issue #481 and #485 Small Bug Fixes (#486)
Addresses issue #481 - when ids were used for events that did not exist
instead of returning a 404 we were getting a null pointer exception and
returning 500. This doesn't hurt much but makes are logs look scarier
than they need be :-)

Also including a one line fix for issue #485 which was noticed while
inspecting logs. There is a spot in our log lines which was intended to
include the env setting but was instead hard coded to TEST, making it
look like prod was misconfigured.