Removing Vutify From DateTimePicker (#488)
I am making use of a library called Flatpickr to replace the date chooser. It is pretty much a drop in replacement.

I am using their 2.x version which supports vue2. They have a 3.x version for when we switch to vue3. I believe the migration should be painless.

I am also replacing the v-button elements