Add analytics gathering for tag suggestion feedback
Adds new "internal API" endpoint on web portal to receive structured analytics data on suggested tags. When the AI tag suggestor is used, this endpoint will be POSTed to with the suggested tags, the final tags, and the ID of the event. Data is sent to Slack for later analysis