Filtering By DateRange Part II -- Admin Endpoint Normalizaiont (#536)
We have a getAll, only exposed to admins. As is unless you are filtering on tags or category it was near un-usable
the endpoint nearly crashes the server when ran because it tries to just get all the things. This changes it to behave
like it's unauthenticated counter part the only difference being that it won't filter un-verified events.

Now it will be paginated with sane defaults that don't crash the server. You can change those defaults and you can also filter by dateRange.

The intent was that these two endpoints should be mostly the same and indeed I had built out all the service logic to make that happen, but until now hadn't got around to updating the endpoint to properly leverage that :-)